관계 및 알림 리팩토링/기능 추가

See merge request s12-webmobile2-sub1/S12P11A606!151

+++ back-end/src/main/java/com/ssafy/pillme/notification/application/response/NotificationSettingResponse.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @Builder
 public class NotificationSettingResponse {
-    private Integer notificationSettingId;
+    private Long notificationSettingId;
     private LocalTime morning;
     private LocalTime lunch;
     private LocalTime dinner;
